#33: Finish merge, add presets limit
